backend/requirements.txt
# ============================================
# SmartScribe Backend - Requirements.txt
# ============================================
# Installez toutes les d�pendances avec:
# pip install -r requirements.txt
# ============================================

# ============================================
# Framework et API Web
# ============================================
fastapi==0.128.0
uvicorn[standard]==0.40.0
python-multipart==0.0.21
websockets==13.1

# ============================================
# Validation et Configuration
# ============================================
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic-core==2.41.5

# ============================================
# Base de donn�es MongoDB
# ============================================
motor==3.7.1
pymongo==4.16.0

# ============================================
# Cache et T�ches Asynchrones
# ============================================
redis==7.1.0
celery==5.6.2
kombu==5.6.2
billiard==4.2.4

# ============================================
# Authentification et S�curit�
# ============================================
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
cryptography==46.0.3

# ============================================
# Traitement Audio
# ============================================
librosa==0.11.0
soundfile==0.13.1
pydub==0.25.1
noisereduce==3.0.3
soxr==1.0.0

# ============================================
# Traitement Vid�o et Images
# ============================================
opencv-python==4.8.1
opencv-contrib-python==4.8.1

# ============================================
# OCR et Traitement d'Images
# ============================================
pytesseract==0.3.10
Pillow==12.1.0
paddlepaddle==2.5.0
paddleocr==2.7.0

# ============================================
# Export de Documents
# ============================================
reportlab==4.4.7
python-docx==1.2.0

# ============================================
# IA et Machine Learning - Core
# ============================================
openai-whisper==20250625
torch==2.9.1
transformers==4.35.0
accelerate==0.25.0
bitsandbytes==0.41.3
trl==0.7.4
peft==0.7.1
sentencepiece==0.2.0
protobuf==4.25.3

# ============================================
# IA - Fine-tuning (Optionnel mais recommand�)
# ============================================
# D�commentez la ligne suivante si vous voulez faire du fine-tuning
# unsloth[colab-new]==2024.8

# ============================================
# Clients IA Cloud
# ============================================
groq==0.4.1
openai==1.12.0
# anthropic==0.18.1  # Optionnel: d�commentez pour Claude API

# ============================================
# Utilitaires G�n�raux
# ============================================
python-dotenv==1.2.1
requests==2.32.5
aiohttp==3.9.1
email-validator==2.3.0
python-dateutil==2.9.0

# ============================================
# D�pendances Num�riques (requises par ML/Audio)
# ============================================
numpy==2.3.5
scipy==1.17.0

# ============================================
# D�veloppement et Tests
# ============================================
pytest==9.0.2
pytest-asyncio==0.21.1
httpx==0.25.2

# ============================================
# Logging et Affichage
# ============================================
colorama==0.4.6

# ============================================
# NOTES IMPORTANTES
# ============================================
# 
# 1. D�PENDANCES SYST�ME (� installer s�par�ment):
#    - FFmpeg: https://ffmpeg.org/download.html
#      Windows: winget install ffmpeg
#    - Tesseract OCR: https://github.com/UB-Mannheim/tesseract/wiki
#      Installez avec support fran�ais
#    - MongoDB: https://www.mongodb.com/try/download/community
#      Windows: winget install MongoDB.Server
#    - Redis: https://github.com/microsoftarchive/redis/releases
#      Windows: winget install Redis.Redis
#    - Ollama (optionnel): https://ollama.ai/download
#      Windows: winget install Ollama.Ollama
#
# 2. PYTORCH AVEC CUDA (si vous avez une carte NVIDIA):
#    Apr�s avoir install� requirements.txt, installez PyTorch avec CUDA:
#    pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118
#    (Remplacez cu118 par votre version CUDA: cu121, cu118, etc.)
#
# 3. PYTORCH CPU (si pas de GPU):
#    Le torch==2.9.1 dans ce fichier est la version CPU par d�faut.
#    Si vous voulez explicitement la version CPU:
#    pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cpu
#
# 4. PADDLEOCR:
#    PaddleOCR n�cessite beaucoup d'espace disque (plusieurs GB).
#    Si vous ne voulez pas l'utiliser, commentez les lignes:
#    # paddlepaddle==2.5.0
#    # paddleocr==2.7.0
#
# 5. UNSLOTH (Fine-tuning):
#    D�commentez la ligne unsloth si vous voulez faire du fine-tuning.
#    N�cessite beaucoup de RAM (16GB+ recommand�).
#
# 6. V�RIFICATION APR�S INSTALLATION:
#    Cr�ez un fichier check_installation.py et ex�cutez-le pour v�rifier
#    que toutes les d�pendances sont correctement install�es.
#
# ============================================
